+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 
+## [4.0.2] - 28.05.2022
+### Changed
+- Added an option to autoplay videos
+- Added an option to disable fullscreen in video
+
 ## [4.0.1] - 10.04.2022
 ### Changed
 - Added the ability to show the number of tweet replies on each tweet